requirements.txt
pyspark==3.5.4
mlflow-skinny==2.15.1
prophet==1.1.5